src/exploration/requirements.txt
certifi==2019.11.28
cycler==0.10.0
decorator==4.4.2
imageio==2.8.0
imutils==0.5.3
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.2.1
networkx==2.4
numpy==1.18.2
opencv-python==4.2.0.32
Pillow==6.2.1
pygame==1.9.6
pyparsing==2.4.6
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
torch==1.1.0
torchvision==0.3.0
